우창하 : 엔진 업데이트

+++ DirectX_MapleStory/GameEngineBase/GameEngineSerializer.cpp
@@ -41,7 +41,7 @@ void GameEngineSerializer::operator<<(const std::string& _Value)
 	// [a][a][a][b][b][b]
 
 	operator<<(static_cast<unsigned int>(_Value.size()));
-	Write(&_Value[0], _Value.size());
+	Write(&_Value[0], static_cast<unsigned int>(_Value.size()));
 }
 void GameEngineSerializer::operator<<(const int& _Value)
 {
